fix: Switch from Sovrin StagingNet to BuilderNet
Due to a recent change, the StagingNet is no longer free. It is
recommended that developers switch to the BuilderNet for development
purposes. If an application reaches a state where it is applicable for a
prototype or pre-production uses, the StagingNet should be used instead.

Please see https://selfserve.sovrin.org/ for more information about the
different Sovrin networks.

This commit changes all references to the StagingNet to point to the
BuilderNet instead.

Signed-off-by: Colton Wolkins (Indicio work address) <colton@indicio.tech>