fix: Stop creating a session just for send_to_admins
Due to changes within the *Managers in ACA-Py, I have decided that it is
cleaner to not create a ProfileSession when we are calling
send_to_admins when the ConnectionManager requires a base profile
anyways. This has resulted in a cleaner codebase for all the places
where send_to_admins is called.

Signed-off-by: Colton Wolkins (Indicio work address) <colton@indicio.tech>